Reorganized a little bit, trying to stay more in line with other specs
